+++ Src/Extension/Diagnostics/PrototypeWizardBusinessComponentNavigationPolicy.cs
@@ -1,0 +1,9 @@
+namespace GenexusOpenApiBuilder.Extension.Diagnostics;
+
+public static class PrototypeWizardBusinessComponentNavigationPolicy
+{
+    public static bool ShouldRequestEnableOnNext(bool isBusinessComponentReady, bool enableBusinessComponentRequested)
+    {
+        return !isBusinessComponentReady && enableBusinessComponentRequested;
+    }
+}